requirements.txt
Flask==2.3.2
Werkzeug==2.3.4
python-barcode==0.14.0
Pillow==9.5.0
reportlab==3.6.12
